Inventory Management App/Backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
mysqlclient>=2.1
boto3>=1.26
python-dotenv>=1.0
django-cors-headers>=4.0
gunicorn>=21.0
